Replace KOHA_NO_TABLE_LOCKS with KOHA_TESTING
We replaced the KOHA_NO_TABLE_LOCKS environment variable with a more
aptly names KOHA_TESTING variable. This variable will now be used in two
places in the koha code base: The existing case in C4::Circulation where
we are using LOCK calls to prevent a race condition and then in the
Koha::Database module where we perform the database connect.

This closes #167