Issue #245 - Add a light test suite run solution
We have few jenkins' nodes and most of them are small and fail to run the test
suite with the elasticsearch container. The whole test suite is not running using
ktd on a server with 8G of RAM. That's not acceptable and we should provide a
lighter solution.
We should have the light solution running on most of the node, and the whole test
suite only on a couple of (big) nodes.

This patch adds a new docker-compose-light.yml file that does not depend
on the es and selenium containers.
It set LIGHT_TEST_SUITE to tell run.sh to not run www and selenium
tests.

Issue #245 - Fix else if syntax error

Issue #245 - Add missing $

oops...

Issue #245 - fix '\' and remove --recursive

Issue #245 - Add more filepaths to ignore

Also add missing xargs