Issue #247 - Embed Jenkins config build script
We have one script per jobs defined in the Jenkins config.
I am suggesting to embed the script in the ktd repo and modify the jenkins config to
wget the script and execute it.

It will prevent copy/paste error and to have to modify all the jobs when a change is needed.

We can decide to move it out of the ktd repo if it does not belong here.