Do not inherit root user's env
This patch removes the -p parameter for koha-shell when running the
tests. It feels wrong to have it, as we are actually passing all
environment variables explicitly, and others should be defined on the
instance's user .bashrc.

Anyhow, this is just an attempt that can be reverted if it has some
other side effects, but my suspicion is that we will just fix new issues
on .bashrc or adding new env variables