Issue #347: Install git hooks when ktd is starting up
We have some git hooks on the wiki
https://wiki.koha-community.org/wiki/Tips_and_tricks for pre-commit and pre-push.
But developers need to install them manually, and it seems that they often forget
(the pre-commit script on the wiki has been broken for several years and nobody
noticed). I am suggesting to embed them in koha-testing-docker and to install them when
the container is started.

Using a path within the koha src dir we are going to make it available from inside and outside
ktd