Issue #352: Pre-commit hook - Embed modified version of the files in the current commit
We are modifying the Vue files using prettier, but we don't embed the modified
version of the files in the current commit.

We need to take precaution here, we don't want to mess with developers
changes!

This commit is embedding the result of prettier into the current commit
(using git add). However we need to make sure there is no local changes
that are not added to the index yet.
For instance:
1. Edit foo.vue
2. git add foo.vue
3. Edit foo.vue again, that would conflict later
4. git commit

I initially tried to resolve the conflict (using the non-added version
of the file, then running the prettier command again), but that seems
messy and dangerous (at least for now). A saner solution seems to just
abort the commit if this case happens, and ask the developer to stash
their changes.