Issue #386: Remove PO files on the right place
On the Docker image build process, each docker command (e.g. RUN)
generates a new layer on top of which the next ones are built. Pretty
much like Git commits.

Because of that, removing the PO files at the end is not saving space
because the PO files are already stored on a previous layer the users
need to download anyway. We just make them not available.

I moved the removal to the same RUN that would generate a filesystem
with them, using the `&&` trick as we already do for not carrying around
the APT cache, etc.

Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>